+++ pyglotaran_extras/types.py
@@ -2,8 +2,10 @@
 from __future__ import annotations
 
 from pathlib import Path
+from typing import Literal
 from typing import Mapping
 from typing import Sequence
+from typing import TypeAlias
 from typing import Union
 
 import xarray as xr
@@ -15,3 +17,27 @@ ResultLike = Union[
     Result, DatasetConvertible, Mapping[str, DatasetConvertible], Sequence[DatasetConvertible]
 ]
 """Result like data which can be converted to a per dataset mapping."""
+
+
+BuiltinSubPlotLabelFormatFunctionKey: TypeAlias = Literal[
+    "number", "upper_case_letter", "lower_case_letter"
+]
+"""Key supported by ``BuiltinLabelFormatFunctions``."""
+
+SubPlotLabelCoordStrs: TypeAlias = Literal[
+    "figure points",
+    "figure pixels",
+    "figure fraction",
+    "subfigure points",
+    "subfigure pixels",
+    "subfigure fraction",
+    "axes points",
+    "axes pixels",
+    "axes fraction",
+    "data",
+    "polar",
+]
+
+SubPlotLabelCoord: TypeAlias = (
+    SubPlotLabelCoordStrs | tuple[SubPlotLabelCoordStrs, SubPlotLabelCoordStrs]
+)